test: fold unretrieved-lock coverage into the blend load/store tests
Move the five sparse-prefetch lock-lifecycle tests out of the standalone
test_blend_release_unretrieved_locks.py into a new section of
test_blend_load_store_opts.py, placed after the existing "L2: obj_keys
cache lifecycle" and "L2: sparse-prefetch read-lock release after
retrieve" sections so both halves of the lock lifecycle read together.

Adapted to that file's conventions: lmcache imports are function-local as
in the rest of it, the constants are namespaced (_UNRETRIEVED_CHUNK /
_UNRETRIEVED_N_CHUNKS), and SessionManager.DEFAULT_SESSION_TTL as a
default argument is replaced by an explicit _UNRETRIEVED_TTL_NEVER, since
the local-import style rules out a class attribute there. A new
_unretrieved_blend() helper replaces the four repeated BlendModule
constructions.

Also drop the list() snapshot in SessionManager.cleanup_expired: collect
the expired sessions first and delete them in a second pass, so the scan
no longer needs a copy of the dict to delete while iterating it.

The moved tests still bite: stubbing out the destroy_listeners
registration in blend.py fails both
test_session_end_releases_unretrieved_sparse_locks and
test_ttl_cleanup_releases_unretrieved_sparse_locks.
